Use float texture coordinates in the range 0.0-1.0. Any fixes, if necessary, should be applied on loading. In the image code (IMAGEDEF) this did not work, so OLD_TEXTURE_SIZE_FIX is applied during rendering.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@2258 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_common/piefunc.h
@@ -34,7 +34,7 @@
 
 extern UBYTE pie_ByteScale(UBYTE a, UBYTE b) WZ_DECL_CONST;
 extern void pie_TransColouredTriangle( PIEVERTEX *vrt, UDWORD rgb );
-extern void pie_DrawSkybox(float scale, int u, int v, int w, int h);
+extern void pie_DrawSkybox(float scale, float u, float v, float w, float h);
 extern void pie_DrawFogBox(float left, float right, float front, float back, float height, float wider);
 extern void pie_DrawViewingWindow( Vector3i *v, UDWORD x1, UDWORD y1, UDWORD x2, UDWORD y2, UDWORD colour);
